Cleanup.

+++ projects/hipblaslt/tensilelite/Tensile/Source/lib/source/analytical/Utils.cpp
@@ -232,7 +232,7 @@ namespace TensileLite
                                                              bool   print,
                                                              size_t WGM)
         {
-            size_t num_threads = std::max((size_t)1, MT_list.size() / 25);
+            size_t num_threads = std::max(1UL, MT_list.size() / 25);
             num_threads = std::min(num_threads, static_cast<size_t>(omp_get_max_threads()));
             omp_set_num_threads(num_threads);
